containers/register-kubeflow-artifacts/requirements.txt
kubeflow-metadata
pandas
pathlib2
